+++ doc/whats-new.md
@@ -2,6 +2,40 @@ What's new
 ==========
 
 
+0.1.1 (unreleased)
+------------------
+
+### New Features
+
+- Add preferences dialog to GUI (#25)\
+  By [Peter Hill](https://github.com/ZedThree) and [John
+  Omotani](https://github.com/johnomotani)
+
+- Catch errors in HypnotoadGui.run(), allows changing settings and pressing Run
+  button again if there was an error in grid generation (#24)\
+  By [John Omotani](https://github.com/johnomotani)
+
+- Plot wall in gui window (#23)\
+  By [John Omotani](https://github.com/johnomotani)
+
+
+### Bug fixes
+
+- If an empty string is passed to a value in the options table of the GUI,
+  resets the option to its default value. Previously this caused a crash (#25)\
+  By [John Omotani](https://github.com/johnomotani)
+
+
+### Internal changes
+
+- Make options table keys in HypnotoadGui immutable. Also makes the use of
+  Options object in Equilibrium/TokamakEquilibrium more consistent by ensuring
+  self.user_options has only been delegated once from
+  TokamakEquilibrium.default_options (i.e. the push() method has only been used
+  once, otherwise update() or set() are called) (#25)\
+  By [John Omotani](https://github.com/johnomotani)
+
+
 0.1.0 (24th April 2020)
 -----------------------